Changed the ReplaceTrees to use Evremonde's method in: http://tpforums.org/forum/showpost.php?p=25308&postcount=22 Changed some ReadInt32s with a cast to uint to ReadUint32. Fixed Item.cs to read the sprite data as a byte array. Renamed Sprites to Sprite in Item and DatItem.

+++ tibiaapi/Objects/Inventory.cs
@@ -424,7 +424,7 @@ namespace Tibia.Objects
             Item item;
             uint address = Addresses.Player.Slot_Head + 12 * ((uint)s - 1);
             byte count = client.ReadByte(address + Addresses.Player.Distance_Slot_Count);
-            uint id = (uint)client.ReadInt32(address);
+            uint id = client.ReadUInt32(address);
             if (id > 0)
             {
                 item = new Item(client, id, count, "", new ItemLocation(s), true);
